Working Config & Score Keeping with Serial Out
The code appeared to work correctly when stepping through debugger,
however interaction with uart and wifi did not work.

Rolling back to a specific version of the arduino expressif libraries
resolved the issue. However, we are now several versions behind and
some current Arduino library functions are missing.

Fixed some bugs in the Score Keeping logic.